Support for connection via Proxy Server
Supports the 3 environment variables HTTP_PROXY, HTTPS_PROXY & NO_PROXY.
These are standard variables used by Golang to support proxy navigation.
These environment variables can be set in the system for all processes.
But since the rhcd clears out all enviornment variables before it calls
a child process we can't have access to the environment variables set
at the service level. To solve that we are storing the proxy values in
our config file and then setting it in our process environment.
If and when rhcd is fixed to allow for propogation of environment
variables to child processes we can remove information from our
config files.